Add files via upload
This is a mini project on text processing that takes a pair of Ethiopian languages from Semitic family. For 
Ethiopian Semitic language family, the pair of languages could be either Amharic and Tigrigna, or 
Amharic and Geez, or Tigrigna and Geez. For this project we chose Amharic and Tigrigna from these pairs
 to work on the text processing.

The task involves
i. Getting text written in the selected Ethiopian language pairs (parallel text, ie text
written on similar topics for each pair of languages)
ii. text pre-processing to clean the text from punctuation marks and other non-phonetic
characters;
iii. Count the frequency of unique words in each language; and computing word level
overlap for the pair of languages to show the extent to which the pair of languages
share words.
iv. Transcribe Ethiopic text into its phonemes (by applying grapheme to phoneme
conversion using g2p algorithms such as Sera)
v. Conduct statistical analysis to compute phonemic distribution of each language; and
phoneme level overlap of the language pairs.